requirements.txt
fastapi
uvicorn[standard]
pydantic
joblib
scikit-learn
numpy
pandas
opentelemetry-exporter-gcp-trace
